refactor: rename 'pipeline' to 'reporting' domain
