+++ src/reporting/reporters/post-pipeline/console/util.ts
@@ -1,5 +1,5 @@
-import { Warnings } from "../../../config/standard/nsci";
-import { consolePrinter } from "../../../lib/console-printer/index.js";
+import { Warnings } from "../../../../config/standard/nsci";
+import { consolePrinter } from "../../../../lib/console-printer/index.js";
 
 export function printWarnOrError(warningsMode: Warnings) {
   /**
